+++ bot/cogs/lobby.py
@@ -3,6 +3,7 @@
 from discord.ext import commands
 from discord import Status as discord_status
 from logging import getLogger
+from modules import tools
 
 from display import AllStrings as disp
 import modules.config as cfg
@@ -36,7 +37,7 @@ class LobbyCog(commands.Cog, name='lobby'):
 
     @commands.command(aliases=['j'])
     @commands.guild_only()
-    async def join(self, ctx):
+    async def join(self, ctx, *args):
         """ Join queue
         """
         if lobby.get_lobby_len() > cfg.general["lobby_size"]:  # This should not happen EVER
@@ -53,17 +54,27 @@ class LobbyCog(commands.Cog, name='lobby'):
         if len(accs) != 0:
             await disp.CHECK_ACCOUNT.send(ctx, cfg.channels["register"], account_names=accs)
             return
-        if player.is_lobbied:
-            await disp.LB_ALREADY_IN.send(ctx)
-            return
         if player.match:
             await disp.LB_IN_MATCH.send(ctx)
             return
+
+        time = await check_time(ctx, args)
+        if time < 0:
+            return
+
+        if player.is_lobbied:
+            if time == 0:
+                await disp.LB_ALREADY_IN.send(ctx)
+            else:
+                player.lobby_expiration = time
+                await disp.LB_TIMEOUT_OK.send(ctx, names_in_lobby=lobby.get_all_names_in_lobby())
+            return
+
         if lobby.is_lobby_stuck():
             await disp.LB_STUCK_JOIN.send(ctx)
             return
 
-        names = lobby.add_to_lobby(player)
+        names = lobby.add_to_lobby(player, expiration=time)
         await disp.LB_ADDED.send(ctx, names_in_lobby=names)
 
     @commands.command(aliases=['rst'])
@@ -76,11 +87,11 @@ class LobbyCog(commands.Cog, name='lobby'):
             await disp.LB_NOT_IN.send(ctx)
             return
         lobby.reset_timeout(player)
-        await disp.LB_REFRESHED.send(ctx)
+        await disp.LB_REFRESHED.send(ctx, names_in_lobby=lobby.get_all_names_in_lobby())
 
     @commands.command(aliases=['l'])
     @commands.guild_only()
-    async def leave(self, ctx):
+    async def leave(self, ctx, *args):
         """ Leave queue
         """
         player = Player.get(ctx.message.author.id)
@@ -88,9 +99,17 @@ class LobbyCog(commands.Cog, name='lobby'):
             await disp.LB_NOT_IN.send(ctx)
             return
         if player.is_lobbied:
-            lobby.remove_from_lobby(player)
-            await disp.LB_REMOVED.send(ctx, names_in_lobby=lobby.get_all_names_in_lobby())
-            return
+            time = await check_time(ctx, args)
+            if time < 0:
+                return
+            elif time == 0:
+                lobby.remove_from_lobby(player)
+                await disp.LB_REMOVED.send(ctx, names_in_lobby=lobby.get_all_names_in_lobby())
+                return
+            else:
+                player.lobby_expiration = time
+                await disp.LB_TIMEOUT_OK.send(ctx, names_in_lobby=lobby.get_all_names_in_lobby())
+                return
         await disp.LB_NOT_IN.send(ctx)
 
     @commands.command(aliases=['q'])
@@ -110,3 +129,20 @@ class LobbyCog(commands.Cog, name='lobby'):
 
 def setup(client):
     client.add_cog(LobbyCog(client))
+
+
+async def check_time(ctx, args):
+    if args:
+        arg = " ".join(args)
+        time = tools.time_calculator(arg)
+        if time == 0:
+            await disp.LB_TIME_INVALID.send(ctx, arg)
+            return -1
+        if time < 300:
+            await disp.LB_TIME_TOO_SHORT.send(ctx)
+            return -1
+        if time > 7200:
+            await disp.LB_TIME_TOO_LONG.send(ctx)
+            return -1
+        return time
+    return 0