+++ bot/display/strings.py
@@ -48,8 +48,12 @@ class AllStrings(Enum):
     LB_NOTIFY = Message("{} queue is almost full ({}/{}), join to start a match!")
     LB_GET = Message("Restore the lobby with `=lobby restore {}`")
     LB_SAVE = Message("Lobby status saved, will be restored on next restart!")
-    LB_REFRESHED = Message("You have reset your queue timeout!")
+    LB_REFRESHED = Message("You have reset your queue timeout!", embed=embeds.lobby_list)
     LB_REFRESH_NO = Message("You can't do that!")
+    LB_TIME_INVALID = Message("Could not interpret `{}` as a valid duration!")
+    LB_TIME_TOO_SHORT = Message("Minimum timeout is 5 minutes!")
+    LB_TIME_TOO_LONG = Message("Maximum timeout is 2 hours!")
+    LB_TIMEOUT_OK = Message("Lobby timeout updated!", embed=embeds.lobby_list)
 
     PK_OVER = Message("The teams are already made. You can't pick!")
     PK_NO_LOBBIED = Message("You must first queue and wait for a match to begin. Check <#{}>")